requirements/base.txt
Django==4.1.4
psycopg2-binary==2.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
python-decouple==3.6
black==22.12.0
django-simple-history==3.0.0
pytest==7.2.0
pytest-django==4.5.2
mixer==7.2.2
django-filter==22.1